==> notebooks/02_eda_2.ipynb <==
import lightgbm as lgb
from lightgbm import early_stopping, log_evaluation
from sklearn.metrics import roc_auc_score  # or accuracy_score for classification
from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np
import matplotlib as plt
from scipy.stats import pearsonr

train_df = pd.read_parquet("../data/train_clean_v3.parquet")
test_df = pd.read_parquet("../data/test_clean_v3.parquet")
test_df = test_df.reset_index()

train_df.replace([np.inf, -np.inf], np.nan, inplace=True)
test_df.replace([np.inf, -np.inf], np.nan, inplace=True)

nan_cols = train_df.columns[train_df.isna().sum() > 0].tolist()

train_df.drop(columns=nan_cols, inplace=True)
test_df.drop(columns=nan_cols, inplace=True)

constant_cols = [col for col in train_df.columns if train_df[col].nunique() == 1]

train_df.drop(columns=constant_cols, inplace=True)
test_df.drop(columns=constant_cols, inplace=True)

skew_vals = train_df.skew(numeric_only=True)
skewed_cols = skew_vals[skew_vals.abs() > 5].index.tolist()
print(skewed_cols)

# train_df.drop(columns=skewed_cols, inplace=True)
# test_df.drop(columns=skewed_cols, inplace=True)

train_df.fillna(train_df.median(), inplace=True)
test_df.fillna(test_df.median(), inplace=True)

train_df.to_parquet("../data/train_clean_v3.parquet")
test_df.to_parquet("../data/test_clean_v3.parquet")

train_df.describe()
